+++ desktop/packages/mullvad-vpn/src/main/grpc-type-convertions.ts
@@ -404,6 +404,8 @@ function convertFromFeatureIndicator(
       return FeatureIndicator.quic;
     case grpcTypes.FeatureIndicator.LWO:
       return FeatureIndicator.lwo;
+    case grpcTypes.FeatureIndicator.PORT:
+      return FeatureIndicator.port;
   }
 }
 
@@ -668,6 +670,9 @@ function convertFromObfuscationSettings(
       break;
     case grpcTypes.ObfuscationSettings.SelectedObfuscation.LWO:
       selectedObfuscationType = ObfuscationType.lwo;
+      break;
+    case grpcTypes.ObfuscationSettings.SelectedObfuscation.PORT:
+      selectedObfuscationType = ObfuscationType.port;
       break;
   }