+++ desktop/packages/mullvad-vpn/src/renderer/components/AppRouter.tsx
@@ -25,6 +25,7 @@ import StateTriggeredNavigation from './StateTriggeredNavigation';
 import Support from './Support';
 import {
   Account,
+  AntiCensorshipView,
   AppInfoView,
   AppUpgradeView,
   ChangelogView,
@@ -42,7 +43,7 @@ import {
   UdpOverTcpSettingsView,
   UserInterfaceSettingsView,
   VpnSettingsView,
-  WireguardSettingsView,
+  WireguardPortView,
 } from './views';
 
 export default function AppRouter() {
@@ -78,7 +79,6 @@ export default function AppRouter() {
           />
           <Route exact path={RoutePath.multihopSettings} component={MultihopSettingsView} />
           <Route exact path={RoutePath.vpnSettings} component={VpnSettingsView} />
-          <Route exact path={RoutePath.wireguardSettings} component={WireguardSettingsView} />
           <Route exact path={RoutePath.daitaSettings} component={DaitaSettingsView} />
           <Route exact path={RoutePath.udpOverTcp} component={UdpOverTcpSettingsView} />
           <Route exact path={RoutePath.shadowsocks} component={ShadowsocksSettingsView} />
@@ -97,6 +97,8 @@ export default function AppRouter() {
           <Route exact path={RoutePath.changelog} component={ChangelogView} />
           <Route exact path={RoutePath.appUpgrade} component={AppUpgradeView} />
           <Route exact path={RoutePath.manageDevices} component={ManageDevicesView} />
+          <Route exact path={RoutePath.antiCensorship} component={AntiCensorshipView} />
+          <Route exact path={RoutePath.wireguardPort} component={WireguardPortView} />
         </Switch>
       </Focus>
     </>