+++ desktop/packages/mullvad-vpn/src/renderer/components/views/anti-censorship/components/method-setting/components/none-option/NoneOption.tsx
@@ -1,0 +1,11 @@
+import { ObfuscationType } from '../../../../../../../../shared/daemon-rpc-types';
+import { messages } from '../../../../../../../../shared/gettext';
+import { SettingsListbox } from '../../../../../../settings-listbox';
+
+export function NoneOption() {
+  return (
+    <SettingsListbox.BaseOption value={ObfuscationType.off}>
+      {messages.gettext('None')}
+    </SettingsListbox.BaseOption>
+  );
+}