+++ desktop/packages/mullvad-vpn/src/renderer/components/views/udp-over-tcp-settings/UdpOverTcpSettingsView.tsx
@@ -1,22 +1,12 @@
-import styled from 'styled-components';
-
 import { messages } from '../../../../shared/gettext';
 import { useHistory } from '../../../lib/history';
 import { AppNavigationHeader } from '../..';
-import * as Cell from '../../cell';
 import { BackAction } from '../../KeyboardNavigation';
 import { Layout, SettingsContainer } from '../../Layout';
 import { NavigationContainer } from '../../NavigationContainer';
 import { NavigationScrollbars } from '../../NavigationScrollbars';
 import SettingsHeader, { HeaderTitle } from '../../SettingsHeader';
 import { UdpOverTcpPortSetting } from './components';
-
-const StyledContent = styled.div({
-  display: 'flex',
-  flexDirection: 'column',
-  flex: 1,
-  marginBottom: '2px',
-});
 
 export function UdpOverTcpSettingsView() {
   const { pop } = useHistory();
@@ -40,11 +30,7 @@ export function UdpOverTcpSettingsView() {
                 </HeaderTitle>
               </SettingsHeader>
 
-              <StyledContent>
-                <Cell.Group>
-                  <UdpOverTcpPortSetting />
-                </Cell.Group>
-              </StyledContent>
+              <UdpOverTcpPortSetting />
             </NavigationScrollbars>
           </NavigationContainer>
         </SettingsContainer>