+++ desktop/packages/mullvad-vpn/src/shared/daemon-rpc-types.ts
@@ -220,6 +220,7 @@ export enum FeatureIndicator {
   shadowsocks,
   quic,
   lwo,
+  port,
   lanSharing,
   dnsContentBlockers,
   customDns,
@@ -491,6 +492,7 @@ export enum ObfuscationType {
   shadowsocks,
   quic,
   lwo,
+  port,
 }
 
 export type ObfuscationSettings = {