+++ desktop/packages/mullvad-vpn/src/shared/routes.ts
@@ -15,7 +15,6 @@ export enum RoutePath {
   userInterfaceSettings = '/settings/interface',
   multihopSettings = '/settings/multihop',
   vpnSettings = '/settings/vpn',
-  wireguardSettings = '/settings/advanced/wireguard',
   daitaSettings = '/settings/daita',
   udpOverTcp = '/settings/advanced/wireguard/udp-over-tcp',
   shadowsocks = '/settings/advanced/shadowsocks',
@@ -34,4 +33,6 @@ export enum RoutePath {
   changelog = '/settings/changelog',
   appUpgrade = '/settings/app-upgrade',
   manageDevices = '/settings/manage-devices',
+  antiCensorship = '/settings/advanced/anti-censorship',
+  wireguardPort = '/settings/advanced/wireguard-port',
 }