+++ desktop/packages/mullvad-vpn/test/e2e/mocked/feature-indicators/feature-indicators.spec.ts
@@ -60,30 +60,37 @@ const featureIndicatorWithOption: FeatureIndicatorWithOptionTestOption[] = [
   {
     testId: 'UDP over TCP',
     featureIndicator: FeatureIndicator.udp2tcp,
-    route: RoutePath.wireguardSettings,
-    featureIndicatorLabel: 'Obfuscation',
-    option: { name: 'Obfuscation', type: 'listbox' },
+    route: RoutePath.antiCensorship,
+    featureIndicatorLabel: 'UDP-over-TCP',
+    option: { name: 'Method', type: 'listbox' },
   },
   {
     testId: 'shadowsocks',
     featureIndicator: FeatureIndicator.shadowsocks,
-    route: RoutePath.wireguardSettings,
-    featureIndicatorLabel: 'Obfuscation',
-    option: { name: 'Obfuscation', type: 'listbox' },
+    route: RoutePath.antiCensorship,
+    featureIndicatorLabel: 'Shadowsocks',
+    option: { name: 'Method', type: 'listbox' },
   },
   {
     testId: 'QUIC',
     featureIndicator: FeatureIndicator.quic,
-    route: RoutePath.wireguardSettings,
-    featureIndicatorLabel: 'Obfuscation',
-    option: { name: 'Obfuscation', type: 'listbox' },
+    route: RoutePath.antiCensorship,
+    featureIndicatorLabel: 'QUIC',
+    option: { name: 'Method', type: 'listbox' },
   },
   {
     testId: 'LWO',
     featureIndicator: FeatureIndicator.lwo,
-    route: RoutePath.wireguardSettings,
-    featureIndicatorLabel: 'Obfuscation',
-    option: { name: 'Obfuscation', type: 'listbox' },
+    route: RoutePath.antiCensorship,
+    featureIndicatorLabel: 'LWO',
+    option: { name: 'Method', type: 'listbox' },
+  },
+  {
+    testId: 'WireGuard Port',
+    featureIndicator: FeatureIndicator.port,
+    route: RoutePath.antiCensorship,
+    featureIndicatorLabel: 'WireGuard Port',
+    option: { name: 'Method', type: 'listbox' },
   },
   {
     testId: 'multihop',
@@ -102,7 +109,7 @@ const featureIndicatorWithOption: FeatureIndicatorWithOptionTestOption[] = [
   {
     testId: 'MTU',
     featureIndicator: FeatureIndicator.customMtu,
-    route: RoutePath.wireguardSettings,
+    route: RoutePath.vpnSettings,
     featureIndicatorLabel: 'MTU',
     option: { name: 'MTU', type: 'input' },
   },
@@ -123,7 +130,7 @@ const featureIndicatorWithOption: FeatureIndicatorWithOptionTestOption[] = [
   {
     testId: 'quantum resistance',
     featureIndicator: FeatureIndicator.quantumResistance,
-    route: RoutePath.wireguardSettings,
+    route: RoutePath.vpnSettings,
     featureIndicatorLabel: 'Quantum resistance',
     option: { name: 'Quantum-resistant tunnel', type: 'switch' },
   },
@@ -147,7 +154,7 @@ test.describe('Feature indicators', () => {
 
   test.beforeEach(async () => {
     await helpers.disconnect();
-    await routes.wireguardSettings.goBackToRoute(RoutePath.main);
+    await routes.antiCensorship.goBackToRoute(RoutePath.main);
   });
 
   test.afterAll(async () => {
@@ -216,7 +223,7 @@ test.describe('Feature indicators', () => {
       'DAITA',
       'Quantum resistance',
       'MTU',
-      'Obfuscation',
+      'UDP-over-TCP',
       'Local network sharing',
       'Lockdown mode',
       'Multihop',