+++ desktop/packages/mullvad-vpn/test/e2e/route-object-models/anti-censorship/selectors.ts
@@ -1,0 +1,17 @@
+import { Page } from 'playwright';
+
+export const createSelectors = (page: Page) => ({
+  wireguardPortButton: () => page.getByRole('button', { name: 'WireGuard Port settings' }),
+  udpOverTcpSettingsButton: () => page.getByRole('button', { name: 'UDP-over-TCP settings' }),
+  udpOverTcpOption: () => page.getByRole('option', { name: 'UDP-over-TCP' }),
+  automaticOption: () =>
+    page
+      .getByRole('listbox', { name: 'Method' })
+      .getByRole('option', { name: 'Automatic', exact: true }),
+  shadowsocksSettingsButton: () => page.getByRole('button', { name: 'Shadowsocks settings' }),
+  shadowsocksOption: () => page.getByRole('option', { name: 'Shadowsocks' }),
+  lwoOption: () => page.getByRole('option', { name: 'LWO' }),
+  quicOption: () => page.getByRole('option', { name: 'QUIC' }),
+  portObfuscationOption: () => page.getByRole('option', { name: 'Port', exact: true }),
+  wireguardPortOption: () => page.getByRole('option', { name: 'WireGuard Port' }),
+});